Add applyMiddleware and compose functions
