+++ kotlin-redux/src/main/kotlin/redux/Imports.kt
@@ -7,11 +7,10 @@ external fun <S, A> createStore(
         enhancer: Enhancer<S, dynamic, A> = definedExternally
 ): Store<S, A>
 
-external interface Store<S, A> {
-    fun getState(): S
-    fun dispatch(action: A): A
-    fun subscribe(listener: () -> Unit): () -> Unit
-    fun replaceReducer(nextReducer: Reducer<S, A>)
-}
+external fun <S, A> combineReducers(reducers: ReducerContainer<S, A>): Reducer<S, A>
 
-external fun <S, A> combineReducers(reducers: dynamic): Reducer<S, A>
+external fun <S, A1, A2> applyMiddleware(middleware: Middleware<S, A1, A2>)
+
+external fun <S, A1, A2, A3> applyMiddleware(middleware1: Middleware<S, A1, A3>, middleware2: Middleware<S, A1, A2>)
+
+external fun compose(vararg functions: dynamic): dynamic
